qcacld-3.0: Search class from global opclass also
While querying for the channel width, the opclass is
extracted from the country opclass. If opclass is not
found then 20Mhz channel width is considered for connection.
This fails the cert test case when AP sends the global
opclass in ECSA.
Change is done to search in the global opclass if the
opclass is not found in the country specific opclass.

CRs-Fixed: 2768075
Change-Id: I24b585af7174a31fe800d6453bbfe18dc5f24af1