qcacld-3.0: Do not allow interop issues ap in FTM mode
Interop issues ap sends vendor event to user space, in FTM
mode vendor events are not initialized which leads to undefined
behavior in kernel if try to send a vendor event if vendor events
are not initialized.

To address this issue, do not allow interop issues ap operation
in FTM mode.

Change-Id: I525dc5ed20a3e615383c535ef65ac7908ede75d4
CRs-Fixed: 3098426