qcacld-3.0: Exempt pe_session validation for set_key_rsp in NAN mode
Currently, host driver validates pe_session in lim to handle
set key response and send the set context response to sme if there
is no valid session. This is fine for STA, P2P and SAP modes as
there would be a valid session. But host driver does not need to
send set context response for NAN iface in NAN pairing protocol
as it doesn't have a valid pe_session and peer does not exist for
NAN iface. So, exempt the pe_session validation for NAN interface.

Change-Id: Iff3ca8b6a281fc0ab94ed8be683ae507e20a214e
CRs-Fixed: 3257689