qcacld-3.0: Make latest driver compatible with existing platform driver
Current platform driver is compatible with host driver
wlan-cld3.driver.lnx.2.0.2.r5 but throws some error when upgraded to
wlan-cld3.driver.lnx.2.0.2.r6. Platform driver can't be updated to latest
version to make it compatible with latest driver as it may break the
existing functionality of the platform.

Resolve the errors from host driver by commenting/removing the unwanted
changes. Most of the changes that are causing errors are from PCIE/MHI,..
which are not needed for x86 platform.

Change-Id: Idb79df437a4215c6fb631b4e52e9235832021503