cld-3.0: Add support for backport flag for he_6ghz_capa parameter
Currently, host driver generates compilation error for
he_6ghz_capa parameter when complies with non-msm kernel 5.7. There is
compilation check for kernel version greater than or equal to 5.4.
But he_6ghz_capa parameter is introduced in station_parameters
structure in non-msm kernel 5.8 but it is backported to msm kernel
5.4.
So, add check for kernel version (greater than or equal to 5.8
or with backported flag, greater than or equal to 5.4)

This change is available on R6 branch only.

Change-Id: I8ae9fe254b44b5e22158e2a74324e600f7bf7ac0