qcacld-3.0: Select Best/Clean channel for SAP
Currently, For CH_WIDTH 20 first clean channel is selected
to start the SAP but logic was not applicable to other band
wdiths because FW can start scan in parallel and for bandwidth
> 20 MHz multiple continuous channels may require to scan and
data was not maintained in host that which channel is scanned
and which is pending.
To Extend logic for bandwidth > 20 MHz, introduce bool array
to maintain data which channel is scanned and found clean. For next
Channel scan same array can be used to check if previous channels
are clean and SAP can be started on current frequency or not.

Change-Id: I9993c0b26885210b4173946e780833dc87830daf
CRs-Fixed: 3403455