qcacld-3.0: handle idle_timeout during SSR
Currently if host driver requests for idle shutdown and SSR is
in progress then platform driver waits for the SSR to complete and
then call the idle shutdown callback to the driver.
But as a part of SSR shutdown sequence, driver flush this idle shutdown
work and this leads to the deadlock.
To resolve this issue, below changes are required.
1) Platform driver rejects the idle shutdown request with -EBUSY
   if SSR is in progress.
2) Host driver runs this idle shutdown work after the reinit if all
   the interfaces are down.
3) Remove the work flush from shutdown sequence.

Change-Id: Icae043971715948280dbafc4023fd58bdc1cc564
CRs-Fixed: 3421819