qcacld-3.0: Handle the error condition in start modules
Currently if cds_dp_open is failed, hdd_start_modules is not
disabling the scheduler thread. So, if we start wlan again,
leads to crash on scheduler restart as it was not
disabled properly.

To fix this add dispatcher_disable in the cleanup in
hdd_start_modules to disable scheduler.

Change-Id: Ibe12e8bf74a8cb8b7604fd2b57426005c922f65a
CRs-Fixed: 3550858