qcacld-3.0: Synchronize country set driver command and idle shutdown
Currently there is no synchronization between country set command
and idle shutdown which leads to undefined behavior in the host
driver.

In the current issue, host driver gets the country change command
and it sends the command to FW, by the time FW sends the response
of the country change command, host driver triggers the idle
shutdown and idle shutdown starts the psoc transition,
in the path to process the regulatory event, host driver tries to
start the psoc operation in hdd country change work handle. Since,
hdd transition is in progress, hdd work handle does not get the
psoc operation and driver reschedules the work until it gets the
psoc operstion. Once the idle shutdown is complete, hdd country
change work handle get the psoc operation object, since idle
shutdown changes the memory domain from active to init domain
and psoc operation gets allocated in init domain. Now by the time
hdd country change work handle executes, host driver got the
interface up and as part of the hdd_open host driver changes the
memory domain from init to active domain. Now, when the hdd country
change work completes, it tries to release the memory of psoc operation,
this gives the issue of incorrect memory domain on mem free because
memory was allocated in init domain and getting released in active
domain.

To fix this issue, take the psoc operation in the host driver when it
receives the country change command so that when idle shutdown is
triggered, it waits for this operation to complete before proceeding
with the psoc transition.
Also, add wait for completion of the country change event in
hdd reg set country API so that psoc operation for country set
gets completed only after the hdd country change work handle is
completed. So that idle shutdown does not proceed and change the
memory domain.
Also, add wait for completion of the country change event in
hdd psoc idle shutdown, so that it does not take the psoc
transition and when kernel schedules hdd country change work,
it get the psoc operation and completes the work.
Since, there are multiple places where host driver is waiting
for country change event in different threads, there is a
possibility that these waits happens in parallel threads.
To complete these all waits, use qdf_event_set_all API instead
of qdf_event_set API to complete the work.

Change-Id: I94c980712f771a4113b83f4a4781072620f68fd1
CRs-Fixed: 3577595