qcacld-3.0: Don't delete ml_peer as part of pe_session cleanup
Currently, sta_ds corresponds to old pe_session bssid is
cleaned up and new sta_ds is created with new bssid when DUT
roams to a new AP. ml_peer is also cleaned up as part of this.
But the ml_peer corresponds to old connection is already
freed up in wma_delete_all_peers() at the beginning of roam
sync handling.
This is not an issue when roamed ml_peer is different from
the current ml_peer as the attempt to delete the old ml_peer
second time gets rejected. But in self roaming cases,
current ml_peer might be valid for other links post roaming.
So, delete the ml_peer in pe_roam_synch_callback() only
if DUT roams to a non-11BE target as the same is not
cleaned up in wma_delete_all_peers() for non-11BE candidates.

Change-Id: I2f92608b880960328cc1b104a8ab55f08c154c42
CRs-Fixed: 3579053