qcacld-3.0: Update op flag for vdev in the critical update case
Update op flag for vdev with correct category in the MLO SAP case
to let wlan firmware know it has critical update. Otherwise fw may
not compare the beacon template content and some critical update ie
like csaie will not be carried on the per-sta profile in the beacon.

Change-Id: I13d74641fe751102ad76fb1025c9c2199591ca0b
CRs-Fixed: 3723988