qcacld-3.0: Cleanup unwanted RSN parsing in LIM
The RSN IE from the hostapd is validated at HDD/SAP and
the information is stored in crypto. There is no need to
validate or cache the RSN in the LIM.

Therefore, cleanup this caching and validation logic in
LIM and refer to crypto for any RSN related information.

Change-Id: Icaf900e4733da99f0d6b794b5e5368cdaccf6869
CRs-Fixed: 3957304