qcacld-3.0: Get scan phymode for legacy connection
While processing a channel load request, the host
can get the scan phymode only if the feature
flag WLAN_FEATURE_11BE is enabled. This results in
the driver being unable to trigger an RRM scan for
11BE-disabled targets.

The fix is to allow the host to calculate the scan
phymode for non-11BE targets as well.

Change-Id: I275efbe12f3b3dd6337193f006b5a82b65f6bcdc
CRs-Fixed: 3949549