qcacld-3.0: Move LAPB timer handler to soft IRQ context
Running the LAPB timer handler in hard IRQ context and waiting for
srng lock can cause the system to hang in case of race conditions.
This change moves the handler to soft IRQ context and tries for the
lock instead of waiting for it inside the handler to prevent such
issues.

Change-Id: I47512fd04f85f9b706a39652774196477d1cea98
CRs-Fixed: 4046313