qcacld-3.0: Do not send RSO START while the link switch is in progress
Currently, the host behavior is:
1. To process the START BSS request from user space, the host disables
roaming on the associated VDEV and re-enables it after the BSS START
process is completed.
2. If a link switch request comes to a non-associated
VDEV, the host does not send an RSO STOP indication to the firmware for
the associated VDEV. Later, the host sends an RSO STOP while processing
the link switch disconnection.

In the case of ML STA + SAP concurrency, when the host enables roaming
in the firmware after BSS START completion and the link switch fails on
the STA, the host fails to bring up the non-associated VDEV. If the
firmware triggers roaming before the host completes the disconnection
for the link switch failure, roaming with only the associated VDEV
active is not expected. This causes the non-associated VDEV to be used
while resetting the spoof MAC address as part of the roam handoff
sequence, resulting in a firmware assertion.

There is no check in the firmware/host to avoid the RSO START command
while the link switch is in progress.

The fix is to send RSO START after BSS START only if the link switch
is not in progress on the STA.

CRs-Fixed: 4013145
Change-Id: I0d309961baad427a823c0822d22ab5f7efe0f545