qcacld-3.0: Cache puncture BW in session
When EHT bandwidth is not equal to HE band width, driver packs
puncturing bitmap in EHT operation and sent it in beacon template
and association response frame.
But for 5 GHz band, driver checks for VHT operation band width
instead of HE bandwidth and it updates VHT channel bandwidth from
the session band width. In case of puncturing, session band width
is different from punctured band width. So, now when driver
compares EHT BW with HE/VHT BW, it found them as equal and hence
it does not include puncturing bitmap in EHT operation.

So, to fix this, cache the punctured BW in session from VDEV
start channel parameters and update the VHT BW to punctured BW
for SAP mode when puncturing is present.

Change-Id: I4db7bf6cfe0584a7b2b3cf55f99dc45c06025bf9
CRs-Fixed: 3995882