qcacld-3.0: Check NDP peer index before increment
Currently, driver assign the index value from the max NDP peer
variable that is cached in NAN PSOC private object and increment
it after index assignment and after that checks for the max index
value. So, every time when driver tries to cache the NDP MAC
address, it increments the NDP peers before comparing it with
maximum NDP peer value. Thus, driver increments the NDP peer value
to 255 and after that it overflows to zero. Hence, it assigns the
index value to 255 and NDP peer value to 0. This leads to crash in
the driver while accessing array at invalid index (255).

So, to fix this, check the index value before incrementing the
NDP peer value.

Change-Id: I9dab85fe3b4d38177c827401f93c83ab0dad6229
CRs-Fixed: 4018423