qcacld-3.0: Clear sinfo before re-using for MLO aggregated stats
Currently in driver, when aggregated mlo stats are requested, the sinfo
structure is reused to get stats for each ml link. This is causing to
report invalid rate info for the second link as the rate field in sinfo
is not cleared properly.
To fix this issue, clear all fields of sinfo before reuse.

Change-Id: Ib1f7362bd0d654aaab579b62168eb960abea5920
CRs-Fixed: 4024752