qcacld-3.0: Drop scan request if standalone DFS SAP is present
Currently, for non-DBS hardware, the host sends a scan request
to the firmware even if a standalone DFS SAP is present, but
the firmware later rejects the scan request.
Make changes to drop the scan request in the host.

Also, host disables STA SAP SCC on DFS channel if
g_sta_sap_scc_on_dfs_chan is 2 and gEnableDFSMasterCap is 0,
as SAP not allowed to operate on DFS if DFS master capability
not set, disable STA SAP SCC on DFS channel if DFS master
capability is not set.

Change-Id: I467e254019ba8c213aa6576ac00a539728b972f7
CRs-Fixed: 4038513