qcacld-3.0: Process CU for all non-zero BPCC values
Host ignores the CU processing if the current BPCC is same as
new BPCC in beacon or probe response. Driver is supposed to
ignore this only if both current and new BPCC are equal to zero.

Restore to original logic to process same BPCC for non-zero value
of BPCC.

Change-Id: Ief5e2686b55823b85f3ad3b8441e7edc8aac06cd
CRs-Fixed: 4021677