qcacld-3.0: Avoid sending duplicate commands during system suspend
Currently, as part of the system suspend process, the host sends
some commands for the associated vdev twice to the firmware.
Additionally, it unnecessarily sends the OBSS scan disable
command to the firmware even if the OBSS scan was not enabled
previously.

Make fix to prevent sending unnecessary commands to the firmware
during system suspend.

Change-Id: I6691369bff3cb861a3a104887690f6ad1c7fd891
CRs-Fixed: 4046137