qcacld-3.0: Pass the active vdev_id during link switch callback
Currently, during a link switch with concurrency, the driver passes
the target vdev_id and attempts to temporarily delete it. However,
if that vdev_id is not active or the in_use flag is not set, the
driver is unable to delete it, causing the link switch to fail
due to the "don't allow 3rd home channel on same MAC" restriction.

Scenarios:
1. ML STA connects on 6 + 5 + 2.4 GHz
2. P2P GO on 5 GHz with gWlanMccToSccSwitchMode=3 and
   non-DBS target (MCC allowed)
3. TTLM command with all TID mapped to 2.4 GHz
4. Link switch happens between 5 and 2.4 GHz. But since the
   in_use flag is not set for vdev_id: 1, the driver is unable
   to delete it, causing the link switch to fail due to ML STA
   on 6 GHz and P2P GO on 5 GHz (already 2 channels present).

Fix is, driver will temporarily delete the active vdev_id.

CRs-Fixed: 4035069
Change-Id: I351ba60855cfa7eba1722952aa178e6ddbef09c5