qcacld-3.0: Fix peer delay disconnection indication to hostapd
When an EB disconnection is in progress and another
disconnect request is received, the current implementation
returns a success status. This causes the thread to get blocked.

To prevent this, return an ‘already in progress’ status
instead of a success status, indicating that another
disconnection request is already being processed.

Change-Id: I7414a7eab04dc62db5b91a4033fd1f7f340ee597
CRs-Fixed: 4059928