qcacld-3.0: Add Aruba OUI to gActionOUIDisableTWT default value
When a TWT session is ongoing with an Aruba AP, after the service period
termination, no data is sent from the AP even though the STA remains in PM0
state. This can lead to an RX throughput drop, with no data being sent for
40-50ms based on the TWT session settings.

To avoid it, add the Aruba OUI to the gActionOUIDisableTWT default value
as requested by the customer, the firmware will avoid setting up TWT
sessions with Aruba APs.

CRs-Fixed: 4049148
Change-Id: Ia42350aa064ab120ec3ef6180cb6d201b9bfd514