qcacld-3.0: Fix in pre_cac work
When SAP is active and the pre_cac interface is created,
a pre_cac flush is performed after the CAC completes.
At this point, the pre_cac work is assigned functions
that are retained for the next interface activation.
In the second iteration, pre_cac cleanup does not occur
because the existing work functions are retained.
This results in the pre_cac vdev not being cleaned up,
leading to a scan failure due to ongoing MLME processes.

So, clear the pre_cac work functions at the pre_cac
iface up.

Change-Id: I116b957eca7f929422e425ac05ed3810e62f81c5
CRs-Fixed: 4093776