qcacld-3.0: Fix beacon unpack issue for roam candidate
Some IOT APs send 6 GHz channel info in the DS IE,
whereas spec allows only 2.4 and 5 GHz channel info
in this IE. This leads to channel to frequency conversion
failure and therefore the scan entry of the roamed link
doesn't get cached in the driver. This ultimately leads
to roam failure due to absence of the link entry.

To fix this, re-order the channel parsing logic.

Change-Id: I133c43b40dd707701000672fd4be76a37ec69db6
CRs-Fixed: 4110948