qcacld-3.0: Update Channel Width for EHT IE
Currently, the host driver receives VHT IE in the beacon,
but the VHT IE is not present in the association response.

As a result, the host driver does not downgrade the channel
width after the association response,
leading to sending peer association with 80 MHz in HT mode.

This change updates the channel width only when EHT IE is present.
Otherwise, the maximum channel width for HT mode will be limited to 40 MHz.

CRs-Fixed: 4139216
Change-Id: Iaf2d3aee1e69c7a555063a00a3a062a8d34f775d