qcacld-3.0: Set unload in progress before psoc trans start
Currently, on rebooting in some devices, platform tries to remove
the wlan driver and it calls PLD remove function directly without
starting driver transition event. But as part of PLD remove,
it start PSOC transition.
Meanwhile userspace is cleaning up and tries to delete all the
interfaces. So, it sends del virtual interface command to the
driver which come up to the driver with RTNL lock and driver is
waiting for PSOC transition to complete before starting the VDEV
transition sync event complete before processing
the del virtual interface command.
While VDEV transition sync is waiting in one thread. The driver
cleanup is happening in another where it tries to unregister the
net devices. For that, it is waiting to take RTNL lock as other
thread (in which del virtual interface command is executing) has
already taken RTNL lock.
Hence, this causes deadlock situation until there is timeout for
VDEV transition sync event.

So, to fix this, set unload in progress flag before starting PSOC
transition at the time of driver remove. This will reject
the VDEV transition sync during delete virtual interface and
release RTNL lock.

Change-Id: I18a5f8b2f90cd202a10488998cdb22cbfb07ab4c
CRs-Fixed: 4083039