qcacld-3.0: Avoid use after free for wlan crypto key entry ptr
In roam sync key event handler, the wlan crypto key pointer
is tried to be mem set to 0 in case of successful and
failed extraction of the keys of individual ML links.
In failure cases such as invalid key index was received
for any of the ML link keys the extract API already frees
the allocated key entries and since the caller also tries
to clear that freed buffer, use after free condition occurs.

Fix is to add NULL pointer validation to avoid the use-after-free.

CRs-Fixed: 4154071
Change-Id: I1a94980855d1db14dc633efd6f51d111bfd29cd7