qcacld-3.0: Skip inactive ML vdev freq for XPAN
In the DBS platform, the STA ML active links can change without
the host's knowledge if the host has indicated number of active
link, as the active link change event is not always communicated
to the host.

To avoid CSA on XPAN when the STA switches active links, the
inactive frequency should be excluded from the XPAN frequency.

CRs-Fixed: 4151363
Change-Id: I606f81c1bfa8a1e575bcbcf3bdd46cf63e4ca592