qcacld-3.0: Stop TDLS timers when STA comes up
Fix TDLS timer handling when STA interface is
brought up after a P2P CLI connection with active
TDLS. Previously, the driver disabled TDLS on the
P2P vdev but did not stop the associated timers,
leading to potential issues.

Fix is, uUpdate the logic to stop TDLS timers
along with disabling TDLS when the STA is
initialized.

CRs-Fixed: 4157033
Change-Id: I59ff65bd1b2c12b6bde0d44ac8c89497628d4fb1