qcacld-3.0: Skip roam enable when concurrent SAP CSA is pending
Currently, STA can connect when SAP is in the middle of CSA and
roaming is also enabled as part of connect procedure. But this
roam enable allow fw to trigger a roaming and that competes with
on-going SAP CSA. This might cause to SAP CSA rejection by
firmware as roaming is in progress and leads to SAP turn-off.

Same is the case when SAP is already present on some channel and
STA gets connected on another channel on the same MAC. SAP would
move to STA channel once STA connection is completed but there is
gap between STA roam enable and SAP decision to move from
MCC to SCC. If firmware triggers roaming before SAP decides to
move(SAP attempts to disable the roaming as part of it), SAP fails
to disable STA roaming and causes contention between STA roaming
and SAP CSA.

Skip roam enable(RSO_ENABLE) when concurrent SAP CSA is pending or
in progress. Roaming gets re-enabled fine from
wlan_hdd_set_roaming_state() on the STA once SAP CSA is done

Change-Id: Ie46ac63fe84607c47024dba60e481ff7ac306d8d
CRs-Fixed: 4174210