qcacld-3.0: Support configuring a set of rates for two types of data streams
To support configuring a set of rates for two types of data
streams for a specified peer. We distinguish different data
streams based on DSCP, each DSCP maps to a TID, and for each
TID, users can configure a set of rate code and one retry
count.

Maximum configurable number of peer is 1.
Maximum configurable number of tid is 2.
Maximum configurable number of rate code is 28.

Change-Id: I520bd2609951842c202e3a8ee0ffe6230ec1ec7a
CRs-Fixed: 4280666