qcacld-3.0: Allow reassoc command if input channel is zero
Recently, a check was added in HDD to reject fast reassoc
command from userspace if input channel frequency was incorrect.
But channel input as zero is valid and should invoke roam,
as per legacy design.
Hence, modify the check to allow the reassoc command if input
channel frequency is zero.

CRs-Fixed: 4128299
Change-Id: Ia27f54a78f586df7262e8ab080dabfe6a0077e86