qcacld-3.0: Do not enable TDLS during link switch
Scenario:
1) The hardware is DBS, with P2P-CLI connection and
3-link MLO STA.
2) The TDLS is enabled for the 2 STA vdevs. As part
of the MLO STA connection, P2P-TDLS gets disabled.
3) Only STA TDLS is allowed during concurrency.
4) Now, Link switch occurs on the assoc vdev, but due
to MCC, the partner vdev is moved to disabled table.
5) After link switch disconnect for assoc vdev, only
the P2P vdev exists on the PM, and the TDLS for P2P
gets enabled.
6) Now, after link switch complete, the TDLS is enabled
for VDEV-0 and later on for VDEV-1 as well.

FW supports a maximum of only 2 TDLS handles, therefore
the 3rd TDLS SET STATE command would get rejected. If
there is any peer addition for that rejected vdev, then
FW enters an undesired state.

To fix this, do not enable any concurrent VDEV TDLS during
link switch. After link switch complete, the TDLS would
be enabled for the ML-STA VDEVs if present.

Change-Id: Idc85364f23b0ceacea48ee57f23154a73ded053d
CRs-Fixed: 4316277